Small changes to allow the names of the variables needed by OMPD to be listed.
Small changes to allow the names of the set of variables
that the OMPD may need to satisfy requests from the debugger.
The file generated by the changes to the CMakefile is not
generated by default, and is not built into the runtime.
So there should be absolutely no performance impact at all.
What it does do is provide a convenient way to expand out
the names that the debugger will eventually be asked about.

  Changes to be committed:
    modified:   runtime/src/CMakeLists.txt
    new file:   runtime/src/ompd-symbols.c.in